Specialize rcu_head_delete for empty deleters.
Empty deleter should take no space.
This is probably still the wrong way to do it; probably rcu_head_delete
should HAVE-A rcu_head instead of BEING-A rcu_head, and then the deleter
should be stored next to it in something like a compressed_pair; but
that's something we can leave to the library vendors, I think.